% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_affinity.R
\name{graph_affinity}
\alias{graph_affinity}
\title{Visualize a Gene's Mutant Peptide - HLA Allele Affinities}
\usage{
graph_affinity(geneName, seq, colorFlag = 0)
}
\arguments{
\item{geneName}{The string name of the gene.}

\item{seq}{The string peptide sequence associated with the gene.}

\item{colorFlag}{By Default, colorFlag is set to 0. If it is set to 1, all
subsequent affinity barplots will be coloured according to binding strength
between the mutation and the HLA-Allele. If it is set to 2, all subsequent
affinity barplots will be coloured according to confidence level between
mutation and the HLA-Allele.}
}
\description{
Barplot a gene's predicted percentage ranks for its present mutant neoantigen
peptides (which have HLA alleles affinities information available in the lung
cancer neoantigen database.)
}
\examples{
graph_affinity("TP53", "FSDLWC....", 1)
}
