% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_affinity_solution}
\alias{graph_affinity_solution}
\title{Produced graph_affinity_solution}
\format{
NULL

 @examples
}
\usage{
graph_affinity_solution
}
\description{
The solution to running graph_affinity_solution, graph_affinity("TP53", tp53mutated3times, 1)
}
\keyword{datasets}
