% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lungData}
\alias{lungData}
\title{Known Lung Cancer Related Neoantigen Database}
\format{

}
\source{
\url{http://www.biostatistics.online/dbPepNeo/index.php/}
}
\usage{
lungData
}
\description{
A dataset containing information pertaining to known neoantigens associated
with lung cancer.
}
\keyword{datasets}
