% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_gene_data.R
\name{mutation_gene_data}
\alias{mutation_gene_data}
\title{Builds a Neoantigen Information Data.Frame Function}
\usage{
mutation_gene_data(geneName, seq)
}
\arguments{
\item{geneName}{The string name of the gene.}

\item{seq}{The string peptide sequence associated with the gene.}
}
\value{
data.frame
}
\description{
Analyzes seq for mutations in the lung cancer neoantigen database for
neoantigens for pertaining to geneName, and returns a data.frame with the
information pertaining to each neoantigen discovered, as well as the
position in seq where the mutation was found.
}
\examples{
mutation_gene_data("KRAS", "AKSVTCTY...")
}
