% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutation_gene_data_solution}
\alias{mutation_gene_data_solution}
\title{Produced mutation_gene_data_solution}
\format{
NULL

 @examples
}
\usage{
mutation_gene_data_solution
}
\description{
The solution to running mutation_gene_data("TP53", tp53mutated3times)
}
\keyword{datasets}
