% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_positions.R
\name{mutation_positions}
\alias{mutation_positions}
\title{The Mutation Position Data.Frame Function}
\usage{
mutation_positions(geneName, seq)
}
\arguments{
\item{geneName}{The string name of the gene.}

\item{seq}{The string peptide sequence associated with the gene.}
}
\value{
data.frame
}
\description{
Analyzes seq for mutations in the lung cancer neoantigen database, for
neoantigens for pertaining to geneName, and returns a data.frame with the
information pertaining to where each neoantigen mutation was positioned in
seq, as well as where the mutation's length.
}
\examples{
mutation_positions("ADGRA3", "TCAKSVTCTY...")
}
