% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutation_positions_solution}
\alias{mutation_positions_solution}
\title{Produced mutation_positions_solution}
\format{
NULL

 @examples
}
\usage{
mutation_positions_solution
}
\description{
The solution to mutation_positions("TP53", tp53mutated3times)
}
\keyword{datasets}
