% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_affinity_visualizer.R
\name{patient_affinity_visualizer}
\alias{patient_affinity_visualizer}
\title{Visualize a Patient's Mutant Peptide - HLA Allele Affinities}
\usage{
patient_affinity_visualizer(geneSet, colNum = 1, colorFlag = 0)
}
\arguments{
\item{geneSet}{A data.frame of gene names and associated gene peptide
sequence.}

\item{colNum}{Number of Columns the resulting set of barplots will have}

\item{colorFlag}{By Default, colorFlag is set to 0. If it is set to 1,
all subsequent affinity barplots will be coloured according to binding
strength between the mutation and the HLA-Allele. If it is set to 2,
all subsequent affinity barplots will be coloured according to confidence
level between mutation and the HLA-Allele.}
}
\value{
status, 1 for failure, 0 for success.
}
\description{
Barplot a set of gene's predicted percentage ranks for their present mutant
neoantigen peptides (which have HLA alleles affinities information available
in the lung cancer neoantigen database.)
}
\examples{
NULL
}
