% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_location_visualizer.R
\name{patient_location_visualizer}
\alias{patient_location_visualizer}
\title{Visualize the Patient's Peptide Locations/Affinities for Neoantigens}
\usage{
patient_location_visualizer(geneSet, colNum = 1, confidence = 0)
}
\arguments{
\item{geneSet}{A data.frame of gene names and associated gene peptide
sequence.}

\item{colNum}{The number of columns you wish the scatterplots to be in.}

\item{confidence}{By Default, confidence is set to 0. If it is set to 1, all
subsequent affinity scatterplots will be coloured according to the
confidence level of the neoantigen.}
}
\value{
status, 1 for failure, 0 for success.
}
\description{
Produce a graphical output portraying a set of gene's neoantigen positions
within their respected genes, as well as predicted percentage ranks for their
present mutant neoantigen peptides (which have HLA alleles affinities
information available in the lung cancer neoantigen database.)
}
\examples{
patient_location_visualizer(testingGeneSet, colNum=1, confidence=0)
}
