% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patient_location_visualizer_solution}
\alias{patient_location_visualizer_solution}
\title{Produced patient_location_visualizer_solution}
\format{
NULL

 @examples
}
\usage{
patient_location_visualizer_solution
}
\description{
The solution to patient_location_visualizer(testingGeneSet, colNum=1, confidence=0)
}
\keyword{datasets}
