% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_neoantigen_list.R
\name{patient_neoantigen_list}
\alias{patient_neoantigen_list}
\title{The Gene Neoantigen Index Function}
\usage{
patient_neoantigen_list(geneSet)
}
\arguments{
\item{geneSet}{A data.frame of gene names and associated gene peptide
sequences.}
}
\value{
A list of indexes
}
\description{
Produces a list of valid indexes into geneSet where a neoantigen target
for lung cancer was discovered. This serves to effectively remove all the
genes from query that do not have neoantigen target information for lung
cancer available.
}
\examples{
NULL
}
