% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutation_location.R
\name{plot_mutation_location}
\alias{plot_mutation_location}
\title{Visualize the Gene's Peptide Locations/Affinities for Neoantigens}
\usage{
plot_mutation_location(geneName, seq, confidence = 0)
}
\arguments{
\item{geneName}{The string name of the gene.}

\item{seq}{The string peptide sequence associated with the gene.}

\item{confidence}{By default, confidence is set to 0. If it is set to 1,
scatterplot will be coloured according to the confidence level of the
neoantigens.}
}
\description{
Produce a graphical output portraying a geneName's neoantigen positions, as
well as predicted percentage ranks for their present mutant neoantigen
peptides (which have HLA alleles affinities information available in the
lung cancer neoantigen database.)
}
\examples{
plot_mutation_location("EPB41L3", "RAMAIYKQSQ...", confidence=0)
}
