% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_mutation_location_solution}
\alias{plot_mutation_location_solution}
\title{Produced plot_mutation_location_solution}
\format{
NULL

 @examples
}
\usage{
plot_mutation_location_solution
}
\description{
The solution to plot_mutation_location("TP53", tp53mutated3times, confidence = 1)
}
\keyword{datasets}
