% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RANBP9mutated}
\alias{RANBP9mutated}
\title{RANBP9 Mutated Peptides}
\format{
A string

 @examples
}
\source{
\url{https://www.uniprot.org/uniprot/Q96S59}
}
\usage{
RANBP9mutated
}
\description{
A string of nucleotides for RANBP9 that has been modified to include
neoantigen relevant mutations.
}
\keyword{datasets}
