% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runlanpAnalysis.R
\name{runlanpAnalysis}
\alias{runlanpAnalysis}
\title{Shiny Applet for lanpAnalysis}
\usage{
runlanpAnalysis()
}
\value{
Return Nothing, but instantiates the shiny applet.
}
\description{
A function that instantiates the shiny applet made for lanpAnalysis
}
\examples{
\dontrun{
runlanpAnalysis()
}

}
