% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solution_patient_neoantigen_list}
\alias{solution_patient_neoantigen_list}
\title{Produced solution_patient_neoantigen_list}
\format{
NULL

 @examples
}
\usage{
solution_patient_neoantigen_list
}
\description{
The solution to patient_neoantigen_list(testingGeneSet)
}
\keyword{datasets}
