% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrounding_peptides.R
\name{surrounding_peptides}
\alias{surrounding_peptides}
\title{The Surrounding Peptide Function}
\usage{
surrounding_peptides(geneName, seq, sides = 3)
}
\arguments{
\item{geneName}{The string name of the gene.}

\item{seq}{The string peptide sequence associated with the gene.}

\item{sides}{The number of peptides you wish to view flanking the ends of
seq. By default this is 3.}
}
\value{
string
}
\description{
Often, the immediate peptides to a mutation are relevant for analyzing a
neoantigen. surrounding_peptides finds all neoantigens within seq for
geneName that are in the lung cancer database, and returns a dataframe
containing the strings for each of these found neoantigens with sides number
of flanking peptides in the peptide sequence (in lower case), and the
peptides' position.
}
\examples{
surrounding_peptides("MAGI2", "GGMNLRIAK...", sides=4)
"(10)rimkNETPVAVLTImiga(26)"
}
