% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surrounding_peptides_solution}
\alias{surrounding_peptides_solution}
\title{Produced surrounding_peptides_solution'}
\format{
NULL

 @examples
}
\usage{
surrounding_peptides_solution
}
\description{
The solution to surrounding_peptides("TP53", tp53mutated3times, 3)
}
\keyword{datasets}
