% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tp53mutated3times}
\alias{tp53mutated3times}
\title{tp53mutated3times Mutated Peptides}
\format{
A string

 @examples
}
\source{
\url{https://www.uniprot.org/uniprot/E7EQX7#sequences}
}
\usage{
tp53mutated3times
}
\description{
A string of nucleotides for tp53mutated3times that has been modified to include
neoantigen relevant mutations in at least 3 locations.
}
\keyword{datasets}
